// src/types.d.ts
export interface IPickInfo<T extends Instance> {
    readonly interaction?: IMouseInteraction | ITouchInteraction;
    readonly layer: string;
    readonly instances: T[];
    readonly screen: [number, number];
    readonly world: [number, number];
    readonly projection: IProjection;
}
export interface IMousePickInfo<T extends Instance> extends IPickInfo<T> {
    readonly interaction: IMouseInteraction;
}
export interface ITouchPickInfo<T extends Instance> extends IPickInfo<T> {
    readonly interaction: ITouchInteraction;
    readonly touch: ISingleTouchInteraction;
}
// ...
export interface IProjection {
    id: string;
    props: IViewProps;
    pixelSpaceToScreen(point: Vec2Compat, out?: Vec2Compat): Vec2Compat;
    screenToPixelSpace(point: Vec2Compat, out?: Vec2Compat): Vec2Compat;
    screenToView(point: Vec2Compat, out?: Vec2Compat): Vec2Compat;
    screenToWorld(point: Vec2Compat, out?: Vec2Compat): Vec2Compat;
    viewToScreen(point: Vec2Compat, out?: Vec2Compat): Vec2Compat;
    viewToWorld(point: Vec2Compat, out?: Vec2Compat): Vec2Compat;
    worldToScreen(point: Vec2Compat, out?: Vec2Compat): Vec2Compat;
    worldToView(point: Vec2Compat, out?: Vec2Compat): Vec2Compat;
}
// ...
export declare enum PickType {
    NONE = 0,
    SINGLE = 1
}
export interface IPickingMetrics {
    currentPickMode: PickType;
    type: PickType;
}
export interface ISinglePickingMetrics<T extends Instance> extends IPickingMetrics {
    type: PickType.SINGLE;
    uidToInstance: Map<number, T>;
}
export interface INonePickingMetrics extends IPickingMetrics {
    type: PickType.NONE;
}
export interface IColorPickingData {
    mouse: Vec2;
    colorData: Uint8Array;
    dataHeight: number;
    dataWidth: number;
    nearestColor: number;
    nearestColorBytes: Vec4;
    allColors: number[];
}